Reorganized controllers.
Moved the modal controllers into their own file; also renamed `SimpDnD4eCtrl` to `DnD4ePageCtrl` since it's job has changed to mostly being a controller for the page and the 'Simple DnD4e' system was renamed a while ago. Also renamed the file to bring it inline with the rest of the files.